refact for use architecture in modules.
